lots of code cleanup, mainly cleanedup the server.go file
